+++ parser/parser.go
@@ -1,21 +1,51 @@
 package parser
 
 import (
-	"encoding/json"
-	"strconv"
 	"time"
 
 	"github.com/twitchscience/spade/reporter"
 )
 
-type MixpanelEvent struct {
-	Pstart     time.Time         // the time that we started processing
-	EventTime  json.Number       // the time that the server recieved the event
-	UUID       string            // UUID of the event as assigned by the edge
-	ClientIp   string            // the ipv4 of the client
-	Event      string            // the type of the event
-	Properties json.RawMessage   // the raw bytes of the json properties sub object
-	Failure    reporter.FailMode // a flag for failure modes
+var parsers = make(map[string]Parser)
+
+// Register makes Parsers available to parse lines. Each Parser should
+// implement an init() call which registers itself by calling this
+// method. Ensure the init() method is called by importing the parser
+// from whichever entry point makes the most sense. For example, in
+// main.go: import _ "github.com/twitchscience/spade/parser/nginx"
+// ensures that the nginx log parser is available
+func Register(name string, p Parser) {
+	if p == nil {
+		panic("parser: Register parser is nil")
+	}
+	if _, dup := parsers[name]; dup {
+		panic("parser: Register called twice for parser: " + name)
+	}
+	parsers[name] = p
+}
+
+type fanoutParser struct {
+	reporter reporter.Reporter
+}
+
+func (f *fanoutParser) Parse(p Parseable) (events []MixpanelEvent, err error) {
+	if nginx, ok := parsers["nginx"]; ok {
+		events, err = nginx.Parse(p)
+		if err != nil {
+			return
+		}
+		if len(events) > 1 {
+			f.reporter.IncrementExpected(len(events) - 1)
+		}
+		return
+	}
+	panic("parser: nginx parser not found")
+}
+
+func BuildSpadeParser(r reporter.Reporter) Parser {
+	return &fanoutParser{
+		reporter: r,
+	}
 }
 
 type Parseable interface {
@@ -27,55 +57,11 @@ type Parser interface {
 	Parse(Parseable) ([]MixpanelEvent, error)
 }
 
+type URLEscaper interface {
+	QueryUnescape([]byte) ([]byte, error)
+}
+
 type ParseResult interface {
 	UUID() string
 	Time() string
-}
-
-func MakeBadEncodedEvent() *MixpanelEvent {
-	return &MixpanelEvent{
-		Pstart:     time.Now(),
-		EventTime:  json.Number(0),
-		UUID:       "error",
-		ClientIp:   "",
-		Event:      "Unknown",
-		Properties: json.RawMessage{},
-		Failure:    reporter.FAILED_TRANSPORT,
-	}
-}
-
-func MakePanicedEvent(line Parseable) *MixpanelEvent {
-	return &MixpanelEvent{
-		Pstart:     line.StartTime(),
-		EventTime:  json.Number(0),
-		UUID:       "error",
-		ClientIp:   "",
-		Event:      "Unknown",
-		Properties: json.RawMessage(line.Data()),
-		Failure:    reporter.PANICED_IN_PROCESSING,
-	}
-}
-
-func MakeErrorEvent(line Parseable, res ParseResult) *MixpanelEvent {
-	uuid := res.UUID()
-	if uuid == "" || len(matches.UUID) > 64 {
-		uuid = "error"
-	}
-	when := res.Time()
-	if when == "" {
-		when = "0"
-	}
-	t, ok := strconv.Atoi(when)
-	if ok != nil {
-		t = 0
-	}
-	return &MixpanelEvent{
-		Pstart:     line.StartTime(),
-		EventTime:  json.Number(t),
-		UUID:       uuid,
-		ClientIp:   "",
-		Event:      "Unknown",
-		Properties: json.RawMessage{},
-		Failure:    reporter.UNABLE_TO_PARSE_DATA,
-	}
 }
