+++ spade_parse/extras.go
@@ -7,7 +7,6 @@ import (
 	"time"
 
 	"github.com/twitchscience/spade/parser"
-	"github.com/twitchscience/spade/parser/nginx"
 	"github.com/twitchscience/spade/reporter"
 	"github.com/twitchscience/spade/table_config"
 	"github.com/twitchscience/spade/transformer"
@@ -58,7 +57,7 @@ func BuildProcessor(configs map[string][]transformer.RedshiftType) *RequestParse
 		t: transformer.NewRedshiftTransformer(
 			table_config.NewStaticLoader(configs),
 		),
-		parser: nginx.BuildSpadeParser(&dummyReporter{}),
+		parser: parser.BuildSpadeParser(&dummyReporter{}),
 	}
 }
